Undo rename of litActions file
